create enviroment variables for client and server
