+++ Client/clientVariables.py
@@ -1,0 +1,4 @@
+#constansts
+
+SERVER_IP= "127.0.0.1"
+PORT= 8006
