+++ Server/serverVariables.py
@@ -1,0 +1,3 @@
+#constansts
+
+PORT = 8006
